Got most of the assignment done
Still have to combine the styles, organizing it and making sure tags work accordingly.  but after 30 hrs of working on this first half of the assignment 2C, unable to think clearly anymore... and still two quizzes to complete in the next 45 mins... I say fuck it.  FAIL if necessary. Can do correctly just NOT in this time frame!